for issue 3126,polish collect logic
