+++ sentinel-extension/pom.xml
@@ -25,7 +25,7 @@
         <module>sentinel-datasource-eureka</module>
         <module>sentinel-annotation-cdi-interceptor</module>
         <module>sentinel-metric-exporter</module>
-        <module>sentinel-prometheus-exporter</module>
+        <module>sentinel-prometheus-metric-exporter</module>
         <module>sentinel-datasource-opensergo</module>
     </modules>
