--- sentinel-extension/sentinel-prometheus-exporter/src/main/resources/META-INF/services/com.alibaba.csp.sentinel.init.InitFunc
@@ -1,1 +1,0 @@
-com.alibaba.csp.sentinel.prom.PromExporterInit
