--- sentinel-extension/sentinel-prometheus-exporter/src/test/java/com/alibaba/csp/sentinel/prom/metrics/GaugeMetricFamilyTest.java
@@ -1,18 +1,0 @@
-package com.alibaba.csp.sentinel.prom.metrics;
-
-import org.junit.Assert;
-import org.junit.Test;
-
-import java.util.Collections;
-
-public class GaugeMetricFamilyTest {
-
-    @Test
-    public void testGaugeMetricFamily(){
-
-        GaugeMetricFamily metricFamily = new GaugeMetricFamily("appName",
-                "sentinel_metrics", Collections.singletonList("type"));
-        metricFamily.addMetric(Collections.singletonList("rt"), 1.0,System.currentTimeMillis());
-        Assert.assertEquals(metricFamily.samples.get(0).value, 1.0,1e-4);
-    }
-}
