+++ sentinel-extension/sentinel-prometheus-metric-exporter/src/main/resources/META-INF/services/com.alibaba.csp.sentinel.init.InitFunc
@@ -1,0 +1,1 @@
+com.alibaba.csp.sentinel.metric.prom.PromExporterInit
